dupadupa

+++ jpp/4/3/final_ostatnie_do_oddania/ConsoleApp1/ConsoleApp1/DHSetup.cs
@@ -8,8 +8,8 @@ public class DHSetup<T> where T: GF, new()
    private T generateGenerator(T dum)
    {
       Random random = new Random();
-      ulong max = dum.GetCharacteristic();
-      ulong number = random.Next(1, max - 1);
+      int max = (int)dum.GetCharacteristic();
+      int number = random.Next(1, max - 1);
       return (T) new GF(number);
    }
